trader/requirements.txt
Backtesting==0.3.3
beautifulsoup4==4.12.3
blinker==1.9.0
bokeh==3.6.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dbpd==1.1
Flask==3.1.0
fonttools==4.55.2
frozendict==2.4.6
gevent==24.11.1
greenlet==3.1.1
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.3
multitasking==0.0.11
numpy==2.1.3
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pyodbc==5.2.0
pyotp==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyNaCl==1.5.0
PyYAML==6.0.2
requests==2.32.3
robin-stocks==3.2.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
ta-lib==0.5.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket==0.2.1
Werkzeug==3.1.3
xyzservices==2024.9.0
yfinance==0.2.50
zope.event==5.0
zope.interface==7.2
